Extract dispatchBuild and redispatchQueuedJob out of submissions.ts
`dispatchBuild` starts a round on a job that has none: it mints the per-job
channel credential, decides whether and how a seed reaches the agent, opens the
round with the selected backend, and records what that cost.
`redispatchQueuedJob` is the reaper's one retry when a dispatch died before it
produced a ref, and calls straight into it. Both move to
`creation/dispatch-build.ts` behind `createDispatcher`, with
`reconstructDispatchSpec`, which only `redispatchQueuedJob` uses.

Bodies unchanged: diffed against master with comments and blank lines stripped,
byte-identical at 167 code lines, and `reconstructDispatchSpec` identical too.

One structural change the extraction forced. `dispatchBuild` was a hoisted
function declaration, so it could reference `seedDeliveryFor`, `seedBuild` and
`publishSeedPreview` from a `const` destructured 350 lines further down. A
factory call is evaluated where it sits, so `createDispatcher` is called after
that destructuring instead. Both consumers — `startImprovementRound` and the
`createGameCreator` wiring — run later still, so nothing observes the change;
`tsc` catches it either way, and did.

Also moves `resumeBuild`'s doc comment into resume-build.ts, where #1052 should
have taken it. Left behind, it sat above a factory call describing a function
that is no longer there — the same stale-comment drift #1051 and #1052 each
found and fixed.

submissions.ts: 1917 -> 1693 lines, ceiling resealed. Since #1046 the file has
lost 1830 lines and its last long-lived helper; what remains is route
registration, the status cache, and the wiring that hands these modules to each
other.
